Done with the 7th and final flexbox exercise
Took a peek at the solution and rewrote a few styles.
Apart from the margin I added, the part where the cards are still look the same despite the changes I made from looking at the solutions,
so I guess the problem is the difference between my screen size and that of the people that provided the exercises.